+++ lib/src/domain/notifiers/simple_state_notifier.dart
@@ -3,6 +3,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:q_architecture/src/domain/mixins/simple_notifier_mixin.dart';
 
+@Deprecated('Use SimpleNotifier instead')
 abstract class SimpleStateNotifier<T> extends StateNotifier<T>
     with SimpleNotifierMixin {
   final Ref ref;